fix: Вернуть хэндлеры обратно в internal
