+++ internal/http-server/handler/create-url-json_test.go
@@ -1,6 +1,7 @@
 package handler
 
 import (
+	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
@@ -9,9 +10,10 @@ import (
 
 	"github.com/dsbasko/yandex-go-shortener/internal/config"
 	"github.com/dsbasko/yandex-go-shortener/internal/entities"
+	"github.com/dsbasko/yandex-go-shortener/internal/http-server/middlewares"
 	"github.com/dsbasko/yandex-go-shortener/internal/storage"
 	"github.com/dsbasko/yandex-go-shortener/internal/urls"
-	"github.com/dsbasko/yandex-go-shortener/pkg/http-server/middlewares"
+	"github.com/dsbasko/yandex-go-shortener/pkg/api"
 	"github.com/dsbasko/yandex-go-shortener/pkg/logger"
 	"github.com/dsbasko/yandex-go-shortener/pkg/test"
 	"github.com/go-chi/chi/v5"
@@ -19,7 +21,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestHandler_CreateURLOnceTextPlain(t *testing.T) {
+func TestHandler_CreateURLJSON(t *testing.T) {
 	config.Init() //nolint:errcheck
 	log := logger.NewMock()
 	store := storage.NewMock(t)
@@ -29,7 +31,7 @@ func TestHandler_CreateURLOnceTextPlain(t *testing.T) {
 	mw := middlewares.New(log)
 	router.
 		With(mw.JWT).
-		Post("/", h.CreateURLTextPlain)
+		Post("/api/shorten", h.CreateURLJSON)
 	ts := httptest.NewServer(router)
 	defer ts.Close()
 
@@ -37,13 +39,22 @@ func TestHandler_CreateURLOnceTextPlain(t *testing.T) {
 
 	tests := []struct {
 		name           string
+		contentType    string
 		body           func() []byte
 		storeCfg       func()
 		wantStatusCode int
 		wantBody       func() string
 	}{
 		{
+			name:           "Wrong Content-Type",
+			body:           func() []byte { return []byte("") },
+			storeCfg:       func() {},
+			wantStatusCode: http.StatusBadRequest,
+			wantBody:       func() string { return "" },
+		},
+		{
 			name:           "Empty Body",
+			contentType:    "application/json",
 			body:           func() []byte { return []byte("") },
 			storeCfg:       func() {},
 			wantStatusCode: http.StatusBadRequest,
@@ -51,7 +62,11 @@ func TestHandler_CreateURLOnceTextPlain(t *testing.T) {
 		},
 		{
 			name: "Service Error",
-			body: func() []byte { return []byte("https://ya.ru/") },
+			body: func() []byte {
+				dtoBytes, _ := json.Marshal(api.CreateURLRequest{URL: "https://ya.ru/"})
+				return dtoBytes
+			},
+			contentType: "application/json",
 			storeCfg: func() {
 				store.EXPECT().
 					CreateURL(gomock.Any(), gomock.Any()).
@@ -62,46 +77,62 @@ func TestHandler_CreateURLOnceTextPlain(t *testing.T) {
 		},
 		{
 			name: "Success Unique",
-			body: func() []byte { return []byte("https://ya.ru/") },
+			body: func() []byte {
+				dtoBytes, _ := json.Marshal(api.CreateURLRequest{URL: "https://ya.ru/"})
+				return dtoBytes
+			},
+			contentType: "application/json",
 			storeCfg: func() {
 				store.EXPECT().
 					CreateURL(gomock.Any(), gomock.Any()).
 					Return(entities.URL{
 						ID:          "42",
 						ShortURL:    "42",
-						OriginalURL: "https://ya.ru/",
+						OriginalURL: "https://ya2.ru/",
 					}, true, nil)
 			},
 			wantStatusCode: http.StatusCreated,
 			wantBody: func() string {
-				return fmt.Sprintf("%s42", config.GetBaseURL())
+				resBytes, _ := json.Marshal(api.CreateURLResponse{
+					Result: fmt.Sprintf("%s42", config.GetBaseURL()),
+				})
+				return string(resBytes)
 			},
 		},
 		{
 			name: "Success NotUnique",
-			body: func() []byte { return []byte("https://ya.ru/") },
+			body: func() []byte {
+				dtoBytes, _ := json.Marshal(api.CreateURLRequest{URL: "https://ya.ru/"})
+				return dtoBytes
+			},
+			contentType: "application/json",
 			storeCfg: func() {
 				store.EXPECT().
 					CreateURL(gomock.Any(), gomock.Any()).
 					Return(entities.URL{
 						ID:          "42",
 						ShortURL:    "42",
-						OriginalURL: "https://ya.ru/",
+						OriginalURL: "https://ya3.ru/",
 					}, false, nil)
 			},
 			wantStatusCode: http.StatusConflict,
 			wantBody: func() string {
-				return fmt.Sprintf("%s42", config.GetBaseURL())
+				resBytes, _ := json.Marshal(api.CreateURLResponse{
+					Result: fmt.Sprintf("%s42", config.GetBaseURL()),
+				})
+				return string(resBytes)
 			},
 		},
 	}
+
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			tt.storeCfg()
 			resp, body := test.Request(t, ts, &test.RequestArgs{
-				Method: "POST",
-				Path:   "/",
-				Body:   tt.body(),
+				Method:      "POST",
+				Path:        "/api/shorten",
+				ContentType: tt.contentType,
+				Body:        tt.body(),
 			})
 			defer resp.Body.Close()
 
@@ -111,7 +142,7 @@ func TestHandler_CreateURLOnceTextPlain(t *testing.T) {
 	}
 }
 
-func BenchmarkHandler_CreateURLOnceTextPlain(b *testing.B) {
+func BenchmarkHandler_CreateURLJSON(b *testing.B) {
 	config.Init() //nolint:errcheck
 	log := logger.NewMock()
 	store := storage.NewMock(&testing.T{})
@@ -121,7 +152,7 @@ func BenchmarkHandler_CreateURLOnceTextPlain(b *testing.B) {
 	mw := middlewares.New(log)
 	router.
 		With(mw.JWT).
-		Post("/", h.CreateURLTextPlain)
+		Post("/api/shorten", h.CreateURLJSON)
 	ts := httptest.NewServer(router)
 	defer ts.Close()
 
@@ -146,9 +177,10 @@ func BenchmarkHandler_CreateURLOnceTextPlain(b *testing.B) {
 		b.StartTimer()
 
 		resp, _ := test.Request(&testing.T{}, ts, &test.RequestArgs{
-			Method: "POST",
-			Path:   "/",
-			Body:   []byte("https://ya.ru/"),
+			Method:      "POST",
+			Path:        "/api/shorten",
+			ContentType: "application/json",
+			Body:        []byte(`{"url":"https://ya.ru/"}`),
 		})
 		resp.Body.Close()
 	}
