+++ internal/http-server/handler/ping.go
